Enhance Discovery class infrastructure
The service files contains a 'txt' field, where service-specific
data can be added. Addition of this text record facilitates unique
identification of IBM BMCs in a network.

The idea of this commit is to enhance the bbclass to append a
new text record in the avahi static service files. This text
record is added only for avahi and not slp.

Tested By:

The additional data should be in the form of 'Key=Value', else the
build will be terminated with the corresponding error.

The following possibilities are tested:
1. Key=Value        => True
2. Key=             => False
3. Key=Value|       => True
4. =Value           => False
5. Key=Value=       => False
6. ==               => False

(From meta-phosphor rev: 2d6aedfb652bc0e6c7657dae586349823e3fb26d)

Signed-off-by: asmithakarun <asmithkr@in.ibm.com>
Change-Id: Ie0b8466a10097d7fdbfd0c8849a0c0d7145e3923
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>