Revert "webui-vue: Add a task to find node licences" (#5)
(cherry picked from commit c91a61d1930285192e955b519b65d2f8c8014694)

This is porting of main line commit:
https://gerrit.openbmc.org/plugins/gitiles/openbmc/openbmc/+/c91a61d1930285192e955b519b65d2f8c8014694
This is required for successful upgrade of OpenSSL from 1.x to 3.1.x.

This reverts commit 28d315377d2a902e2d3691159411c18a2ce64a66.

This commit no longer compiles with the latest upstream yocto rebase. 
I've spent a few hours trying to fix it but no luck. Regenerating the
files did not help.

The error is the following:
npm verb stack Error: request to https://registry.npmjs.org/archy
failed: cache mode is 'only-if-cached' but no cached response is
available.

This appears to just be for licensing information so not something
required for the function to work. Revert until a fix can be identified.

Tested:
    - Image compiles fine with this change.

Change-Id: I2863ffada3e8b0124fa0cf5194f4407488f08e15 
Signed-off-by: Saravanan Palanisamy <saravanan.palanisamy@intel.com>
Co-authored-by: Andrew Geissler <geissonator@yahoo.com>